# Optimize leftpad on raw_html