--- a/lib/floki/raw_html.ex
+++ b/lib/floki/raw_html.ex
@@ -43,7 +43,7 @@ defmodule Floki.RawHTML do
 
     padding =
       case Keyword.fetch(options, :pretty) do
-        {:ok, true} -> %{pad: "  ", line_ending: "\n", depth: 0}
+        {:ok, true} -> %{pad: "", pad_increase: "  ", line_ending: "\n", depth: 0}
         _ -> :noop
       end
 
@@ -211,7 +211,7 @@ defmodule Floki.RawHTML do
     do: map_intersperse(Map.to_list(attrs), separator, mapper)
 
   defp leftpad(:noop), do: ""
-  defp leftpad(%{pad: pad, depth: depth}), do: String.duplicate(pad, depth)
+  defp leftpad(%{pad: pad}), do: pad
 
   defp leftpad_content(:noop, string), do: string
 
@@ -226,7 +226,11 @@ defmodule Floki.RawHTML do
   end
 
   defp pad_increase(:noop), do: :noop
-  defp pad_increase(padder = %{depth: depth}), do: %{padder | depth: depth + 1}
+
+  defp pad_increase(padder = %{depth: depth, pad_increase: pad_increase}) do
+    depth = depth + 1
+    %{padder | depth: depth, pad: String.duplicate(pad_increase, depth)}
+  end
 
   defp line_ending(:noop), do: ""
   defp line_ending(%{line_ending: line_ending}), do: line_ending